Submitting solution to Problem 24
This is not my proudest moment. I've been out of the Project Euler game for a while and just wanted to get back into it
without doing much thinking.

I initially attempted this with pen and paper, but kept on getting stuck on something. Yesterday I came back to this and
couldn't find my original notes, hence went "fuck it" and brute forced it.

The initial approach was going to make use of factorials and recursion to work out each digit in the sequence from the
left onwards.

Sometimes I really hate myself.

Issue(s): None